requirements.txt
easyocr>=1.6.0
opencv-python-headless>=4.5.3
numpy>=1.21.0
pyttsx3>=2.90
gTTS>=2.2.3
sounddevice>=0.4.1
soundfile>=0.10.3
Flask>=2.0.0
Pillow>=8.3.1
langdetect>=1.0.9
serverless-wsgi>=3.0.1
werkzeug>=2.0.0 